requirement.txt
anyio==4.6.2.post1
APScheduler==3.6.3
attrs==24.2.0
cachetools==4.2.2
certifi==2024.8.30
charset-normalizer==3.4.0
et_xmlfile==2.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
numpy==2.1.3
openpyxl==3.1.5
outcome==1.3.0.post0
pandas==2.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-telegram-bot==13.15
pytz==2024.2
requests==2.32.3
selenium==4.27.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tornado==6.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
XlsxWriter==3.2.0
